remove upper bound on python_requires (#1071)
Python dependency managers Poetry and PDM (and perhaps others) tend to automatically set an upper bound of 4.0 for the required-python variable. This prevents CTranslate2 from getting installed, without much detail provided on exactly why - especially if CTranslate2 is a dependency of the actual package being installed. This PR removes the upper bound, which tends not to be used by most other packages.