feat: adding configuration to recently opened file enable collection of snippets from recent opened files
This commit enables the collection of snippets from recent opened files in the CompletionProvider class. It adds a new configuration option `collectSnippetsFromRecentOpenedFiles` to the default configuration, which is set to `true` by default. The maximum number of opened files to collect snippets from is set to 5.

These changes are necessary to improve the code completion feature by including snippets from recently opened files.

Ref: feat(recent-opened-files): enable collection of snippets from recent opened files